langchain/requirements.txt
streamlit
python-dotenv
langchain
langchain_core
langchain-google-genai
langserve
fastapi
uvicorn
sse_starlette
langchain_community
bs4
pypdf
langchain_text_splitters
ChromaDb
langchain_google_genai
faiss-cpu
ddgs
langchain_experimental
wikipedia